+++ DatadogInternal/Sources/NetworkInstrumentation/B3/B3HTTPHeadersWriter.swift
@@ -79,17 +79,17 @@ public class B3HTTPHeadersWriter: TracePropagationHeadersWriter {
 
     /// Initializes the headers writer.
     ///
-    /// - Parameter sampleRate: The sampling rate applied for headers injection, with 20% as the default.
+    /// - Parameter sampleRate: The sampling rate applied for headers injection, with 100% as the default.
     /// - Parameter injectEncoding: The B3 header encoding type, with `.single` as the default.
     @available(*, deprecated, message: "This will be removed in future versions of the SDK. Use `init(samplingStrategy: .custom(sampleRate:))` instead.")
     public convenience init(
-        sampleRate: Float = 20,
+        sampleRate: Float = .maxSampleRate,
         injectEncoding: InjectEncoding = .single
     ) {
         self.init(
             samplingStrategy: .custom(sampleRate: sampleRate),
             injectEncoding: injectEncoding,
-            traceContextInjection: .all
+            traceContextInjection: .sampled
         )
     }
 
@@ -97,11 +97,11 @@ public class B3HTTPHeadersWriter: TracePropagationHeadersWriter {
     ///
     /// - Parameter samplingStrategy: The strategy for sampling trace propagation headers.
     /// - Parameter injectEncoding: The B3 header encoding type, with `.single` as the default.
-    /// - Parameter traceContextInjection: The trace context injection strategy, with `.all` as the default.
+    /// - Parameter traceContextInjection: The trace context injection strategy, with `.sampled` as the default.
     public init(
         samplingStrategy: TraceSamplingStrategy,
         injectEncoding: InjectEncoding = .single,
-        traceContextInjection: TraceContextInjection = .all
+        traceContextInjection: TraceContextInjection = .sampled
     ) {
         self.samplingStrategy = samplingStrategy
         self.injectEncoding = injectEncoding