+++ DatadogObjc/Sources/Tracing/Propagation/B3HTTPHeadersWriter+objc.swift
@@ -48,13 +48,13 @@ public class DDB3HTTPHeadersWriter: NSObject {
     @objc
     @available(*, deprecated, message: "This will be removed in future versions of the SDK. Use `init(samplingStrategy: .custom(sampleRate:))` instead.")
     public init(
-        sampleRate: Float = 20,
+        sampleRate: Float = .maxSampleRate,
         injectEncoding: DDInjectEncoding = .single
     ) {
         swiftB3HTTPHeadersWriter = B3HTTPHeadersWriter(
             samplingStrategy: .custom(sampleRate: sampleRate),
             injectEncoding: .init(injectEncoding),
-            traceContextInjection: .all
+            traceContextInjection: .sampled
         )
     }
 
@@ -62,7 +62,7 @@ public class DDB3HTTPHeadersWriter: NSObject {
     public init(
         samplingStrategy: DDTraceSamplingStrategy,
         injectEncoding: DDInjectEncoding = .single,
-        traceContextInjection: DDTraceContextInjection = .all
+        traceContextInjection: DDTraceContextInjection = .sampled
     ) {
         swiftB3HTTPHeadersWriter = B3HTTPHeadersWriter(
             samplingStrategy: samplingStrategy.swiftType,