+++ DatadogRUM/Sources/RUMConfiguration.swift
@@ -352,23 +352,23 @@ extension RUM.Configuration.URLSessionTracking {
         ///
         /// - Parameters:
         ///   - hosts: The set of hosts to inject tracing headers. Note: Hosts must not include the "http(s)://" prefix.
-        ///   - sampleRate: The sampling rate for tracing. Must be a value between `0.0` and `100.0`. Default: `20`.
-        ///   - traceControlInjection: The strategy for injecting trace context into requests. Default: `.all`.
+        ///   - sampleRate: The sampling rate for tracing. Must be a value between `0.0` and `100.0`. Default: `100`.
+        ///   - traceControlInjection: The strategy for injecting trace context into requests. Default: `.sampled`.
         case trace(
             hosts: Set<String>,
-            sampleRate: Float = 20,
-            traceControlInjection: TraceContextInjection = .all
+            sampleRate: Float = .maxSampleRate,
+            traceControlInjection: TraceContextInjection = .sampled
         )
 
         /// Trace given hosts with using custom tracing headers.
         ///
         /// - `hostsWithHeaders` - Dictionary of hosts and tracing header types to use. Note: Hosts must not include "http(s)://" prefix.
-        /// - `sampleRate` - The sampling rate for tracing. Must be a value between `0.0` and `100.0`. Default: `20`.
-        /// - `traceControlInjection` - The strategy for injecting trace context into requests. Default: `.all`.
+        /// - `sampleRate` - The sampling rate for tracing. Must be a value between `0.0` and `100.0`. Default: `100`.
+        /// - `traceControlInjection` - The strategy for injecting trace context into requests. Default: `.sampled`.
         case traceWithHeaders(
             hostsWithHeaders: [String: Set<TracingHeaderType>],
-            sampleRate: Float = 20,
-            traceControlInjection: TraceContextInjection = .all
+            sampleRate: Float = .maxSampleRate,
+            traceControlInjection: TraceContextInjection = .sampled
         )
     }